Update provide dependencies
Remove the obsolete template provider
Update tls provider to version 4.

Upstream changelog notes:

"resource/tls_self_signed_cert: private_key_pem attribute is now stored
in the state as-is; first apply may result in an update-in-place"

This is likely to update the CA certificate in place, and may require a
control plane restart.

Fixes #16